Add Gyro class
