+++ src/main/java/frc/robot/Constants.java
@@ -1,5 +1,5 @@
 package frc.robot;
 
 public class Constants {
-    
+    public final static int FILTER_WINDOW_SIZE = 5;
 }
