--- src/main/java/frc/robot/helpers/Gyro.java
@@ -1,16 +1,0 @@
-package frc.robot.helpers;
-
-import edu.wpi.first.wpilibj.ADIS16448_IMU;
-
-public class Gyro {
-    public ADIS16448_IMU imu = new ADIS16448_IMU();
-
-    public Gyro() {
-        imu.reset();
-    }
-
-    public void periodic() {
-
-    }
-
-}
