+++ src/main/java/frc/robot/subsystems/Drivetrain.java
@@ -7,7 +7,6 @@ import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 
 import edu.wpi.first.wpilibj.drive.DifferentialDrive;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.helpers.Gyro;
 
 public class Drivetrain extends SubsystemBase {
     CANSparkMax frontLeftDrive = new CANSparkMax(1, MotorType.kBrushless);
